Add base test control object
  * Use it in test control for web platform
